requirements.txt
# National Highways Management System
# Python Dependencies

# Database
mysql-connector-python==8.2.0

# Password hashing
bcrypt==4.1.1

# Web framework (optional - for REST API)
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.5.3

# API documentation (optional)
flask-swagger-ui==4.11.1

# Data validation
pydantic==2.5.0

# Production server
gunicorn==21.2.0
werkzeug==3.0.1

# Environment variables
python-dotenv==1.0.0

# Date/time handling
python-dateutil==2.8.2

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development tools
black==23.12.0
flake8==6.1.0
mypy==1.7.1
